create price, save and load price

+++ Assets/_Main/Scripts/Point/PriceUI.cs
@@ -17,7 +17,7 @@ public class PriceUI : BaseMonoBehaviour
         PriceManager.Instance._UpdatePriceUI -= UpdatePriceUI;
     }
 
-    private void UpdatePriceUI(int price)
+    private void UpdatePriceUI(double price)
     {
         _text.text = price.ToString();
     }
